Begin exposing the share links
